Save margin catalog to disk
